# Pre-allocate and reuse value-path for encoding