@fxamacker 2021-04-29
@turbolent I agree we should avoid side effects.  There's a good chance I may have misunderstood the code review.  Please let me know.

If I understand you correctly, calling function's `path` must not be modified by the called function.

Unless I'm mistaken, the calling function's `path` will still have the same element values and number of elements, because the called functions only modify new elements not visible to the calling function when the called function returns.

For example, in Go 1.15.11:
```Go
func m1(path []string) {
	path1 := append(path, "m1")
	fmt.Printf("m1()   path1 %v, len %d, cap %d\n", path1, len(path1), cap(path1))
	path1[1] = "xx"
	fmt.Printf("m1()   path1 %v, len %d, cap %d\n", path1, len(path1), cap(path1))
}

func main() {
	path := []string{"m0"}
	fmt.Printf("main() path  %v, len %d, cap %d\n", path, len(path), cap(path))
	m1(path)
	fmt.Printf("main() path  %v, len %d, cap %d\n", path, len(path), cap(path))
}
```

Produces:
```
main() path  [m0], len 1, cap 1
m1()   path1 [m0 m1], len 2, cap 2
m1()   path1 [m0 xx], len 2, cap 2
main() path  [m0], len 1, cap 1
```
https://play.golang.com/p/RDCcSBs7GVh