@turbolent 2021-05-03
Yes, that's a good summary of the two points!

> 1. Further performance gains are possible by using a different data structure

Yes, we should definitely get this PR in first, I was only thinking out loud and brought up this idea, it was not meant to be addressed in this PR 👍 

> 2. calls to append() in this PR mutate the path argument and would affect the calling code

My main point was that `append` modifies the slice's underlying array, so we have to be careful with this side-effect, e.g. see
- https://blog.golang.org/slices#TOC_4.
- https://www.tugberkugurlu.com/archive/working-with-slices-in-go-golang-understanding-how-append-copy-and-slicing-syntax-work#calling-append-on-a-sliced-slice-may-modify-the-original-slice

For example (https://play.golang.org/p/SsapANm9Q-p):

```go
func use (b []int) {
	fmt.Printf("b: %v, cap: %d\n", b, cap(b))
	b = append(b, 20)
	fmt.Printf("b: %v, cap: %d\n", b, cap(b))
}

func main() {
	a := []int{1, 2, 3, 4, 5}
	fmt.Printf("a: %v, cap: %d\n", a, cap(a))

	use(a[2:4])
	fmt.Printf("a: %v, cap: %d\n", a, cap(a))
}
```

```
a: [1 2 3 4 5], cap: 5
b: [3 4], cap: 3
b: [3 4 20], cap: 3
a: [1 2 3 4 20], cap: 5
```

> this PR won't cause `v=append(path,"")` followed by `v[len(path)]="abc"` to mutate path. 

It does, though, I think in our particular case here, this is not a problem: 
Even though a function call deeper in the call stack modifies the slice by extending it (`append`ing one or more elements) and writing to the newly appended element/index, functions up the call stack won't observe it because the slice header's length won't change, and deeper functions only ever *add*, they always pass the full slice.

For example, if function A calls B, passing the path slice of length 3, and B appends one element and modifies it, then A's slice header still has a length of 3, though the capacity of the slice is now >3 and there's an additional element in the array pointed to by the slice header (not visible in A).

I then think it's fine to keep removing the defensive copies in the "deeper" encoding functions, but we must be very careful that we keep the invariant that all encoding functions only append, not modify parts in the passed path, because those changes are visible to the parent function)