@fxamacker 2021-05-03
> Do we already have an issue for applying the same changes to the decoder?

Great question!  I don't think we do.  We should probably open an issue to assess whether the same changes can be done.